@@ -1,12 +1,14 @@
 {
 	"name": "kaikei",
-	"version": "1.1.0",
+	"version": "2.0.0",
 	"description": "",
 	"main": "index.js",
+	"type": "module",
 	"scripts": {
-		"build-production": "NODE_OPTIONS=--openssl-legacy-provider webpack -c --mode production",
-		"build": "NODE_OPTIONS=--openssl-legacy-provider webpack -c --mode development",
-		"watch": "NODE_OPTIONS=--openssl-legacy-provider webpack --watch",
+		"build": "vite build",
+		"build-ssr": "vite build --config vite.config.ssr.js",
+		"watch": "vite build -w",
+		"watch-ssr": "vite build -w --config vite.config.ssr.js",
 		"prepublishOnly": "npm test",
 		"start-dev": "node ./bin/www",
 		"start": "NODE_ENV='production' node ./bin/www"
@@ -14,47 +16,91 @@
 	"author": "ogochan@wasp.co.jp",
 	"license": "AGPL",
 	"dependencies": {
+		"@codemirror/lang-javascript": "^6.2.4",
+		"@codemirror/lang-markdown": "^6.3.2",
+		"@formkit/tempo": "^0.1.2",
 		"@fortawesome/fontawesome-free": "^6.1.1",
+		"@iconify/svelte": "^4.2.0",
 		"@popperjs/core": "^2.9.2",
-		"axios": "^0.21.1",
+		"@roxi/routify": "^2.18.18",
+		"@sveltejs/vite-plugin-svelte": "^3.1.2",
+		"@tinymce/tinymce-svelte": "^3.0.0",
+		"axios": "^1.7.7",
 		"bcrypt": "^5.0.1",
 		"bootstrap": "^5.0.1",
 		"bootstrap-icons": "^1.5.0",
+		"chart.js": "^4.4.4",
+		"codemirror": "^6.0.1",
 		"connect-multiparty": "^2.2.0",
 		"cookie-parser": "^1.4.5",
 		"cors": "^2.8.5",
+		"dotenv": "^16.5.0",
+		"ejs": "^3.1.10",
 		"exceljs": "^4.2.1",
 		"express": "^4.17.1",
 		"express-session": "^1.17.2",
 		"file-loader": "^6.2.0",
+		"gridstack": "^11.5.0",
+		"jspreadsheet-ce": "^5.0.0",
+		"markdown-it": "^14.1.0",
+		"markdown-it-abbr": "^2.0.0",
+		"markdown-it-anchor": "^9.2.0",
+		"markdown-it-attrs": "^4.2.0",
+		"markdown-it-container": "^4.0.0",
+		"markdown-it-deflist": "^3.0.0",
+		"markdown-it-emoji": "^3.0.0",
+		"markdown-it-footnote": "^4.0.0",
+		"markdown-it-ins": "^4.0.0",
+		"markdown-it-mark": "^4.0.0",
+		"markdown-it-prism": "^2.3.0",
+		"markdown-it-replace-link": "^1.2.2",
+		"markdown-it-sub": "^2.0.0",
+		"markdown-it-sup": "^2.0.0",
 		"mime": "^3.0.0",
+		"mitt": "^3.0.1",
 		"morgan": "^1.10.0",
-		"passport": "^0.5.2",
+		"morphdom": "^2.7.5",
+		"passport": "^0.7.0",
 		"passport-local": "^1.0.0",
+		"path-browserify": "^1.0.1",
 		"pg": "^8.6.0",
 		"pg-hstore": "^2.3.3",
+		"playwright": "^1.51.1",
+		"prismjs": "^1.29.0",
 		"sequelize": "^6.6.2",
 		"sequelize-cli": "^6.2.0",
 		"session-file-store": "^1.5.0",
+		"sortable": "^2.0.0",
+		"sortablejs": "^1.15.6",
 		"sprightly": "^1.0.3",
+		"svelte": "^4.2.19",
+		"svelte-chartjs": "^3.1.5",
+		"svelte-codemirror-editor": "^1.4.1",
+		"svelte-hmr": "^0.16.0",
 		"svelte-loader": "^3.1.2",
-		"svelte-preprocess": "^4.10.4",
+		"svelte-toasts": "^1.1.2",
 		"svgo": "^2.8.0",
 		"svgo-loader": "^3.0.0",
-		"url-loader": "^4.1.1"
+		"tinymce": "^7.5.1",
+		"url-loader": "^4.1.1",
+		"uuid": "^11.1.0"
 	},
 	"devDependencies": {
 		"chai": "^4.2.0",
 		"css-loader": "^5.2.4",
+		"esbuild": "^0.23.1",
 		"esm": "^3.2.25",
-		"jsdom": "^16.2.2",
+		"jsdom": "^16.7.0",
 		"jsdom-global": "^3.0.2",
-		"mocha": "^8.2.1",
+		"mocha": "^10.4.0",
 		"nyc": "^15.1.0",
+		"postcss-url": "^10.1.3",
+		"sass": "^1.77.8",
 		"sequelize-auto": "^0.8.3",
 		"style-loader": "^2.0.0",
-		"webpack": "^4.46.0",
-		"webpack-cli": "^3.3.12",
-		"webpack-dev-server": "^3.11.0"
+		"vite": "^5.4.3",
+		"vite-plugin-static-copy": "^2.2.0",
+		"webpack": "^5.90.3",
+		"webpack-cli": "^5.1.4"
 	}
 }
