@@ -1,8 +1,6 @@
-'use strict';
-const {
-	Model
-} = require('sequelize');
-module.exports = (sequelize, DataTypes) => {
+import {Model} from 'sequelize';
+
+export default (sequelize, DataTypes) => {
 	class AccountRemaining extends Model {
 		/**
 		 * Helper method for defining associations.
