@@ -1,8 +1,6 @@
-'use strict';
-const {
-	Model
-} = require('sequelize');
-module.exports = (sequelize, DataTypes) => {
+import {Model} from 'sequelize';
+
+export default (sequelize, DataTypes) => {
 	class CrossSlip extends Model {
 		/**
 		 * Helper method for defining associations.
@@ -12,8 +10,28 @@ module.exports = (sequelize, DataTypes) => {
 		static associate(models) {
 			this.hasMany(models.CrossSlipDetail, {
 				foreignKey: 'crossSlipId',
-				sourceKey: 'id',
 				as: 'lines'
+			});
+			this.hasOne(models.User, {
+				foreignKey: 'id',
+				sourceKey: 'createdBy',
+				as: 'creater',
+				onDelete: 'RESTRICT',
+				onUpdate: 'CASCADE'
+			});
+			this.hasOne(models.User, {
+				foreignKey: 'id',
+				sourceKey: 'approvedBy',
+				as: 'approver',
+				onDelete: 'RESTRICT',
+				onUpdate: 'CASCADE'
+			});
+			this.hasOne(models.User, {
+				foreignKey: 'id',
+				sourceKey: 'updatedBy',
+				as: 'updater',
+				onDelete: 'RESTRICT',
+				onUpdate: 'CASCADE'
 			});
 		}
 	};
@@ -23,7 +41,11 @@ module.exports = (sequelize, DataTypes) => {
 		day: DataTypes.INTEGER,
 		no: DataTypes.INTEGER,
 		lineCount: DataTypes.INTEGER,
-		term: DataTypes.INTEGER
+		term: DataTypes.INTEGER,
+		approvedAt: DataTypes.DATE,
+		approvedBy: DataTypes.INTEGER,
+		createdBy: DataTypes.INTEGER,
+		updatedBy: DataTypes.INTEGER
 	}, {
 		sequelize,
 		modelName: 'CrossSlip',
