@@ -1,159 +1,233 @@
-const models = require('../models');
+import models from '../models/index.js';
 const Op = models.Sequelize.Op;
+import {create as createCrossSlip, update as updateCrossSlip} from '../libs/cross_slip.js';
 
-module.exports = {
-	get: async(req, res, next) => {
-		let year = req.params.year;
-		let month = req.params.month;
-		let no = req.params.no;
+export default {
+  list: async(req, res, next) => {
+    res.set('Access-Control-Allow-Origin', '*');
+    switch  ( req.params.type)  {
+      case  'not_approved':
+        let where;
+        if  ( req.session.user.approvable ) {
+          where = {
+            approvedAt: {
+              [Op.eq]: null
+            },
+            term: req.session.term
+          }
+        } else {
+          where = {
+            approvedAt: {
+              [Op.eq]: null
+            },
+            term: req.session.term,
+            createdBy: req.session.user.id
+          }
+        }
+        let cross_slips = await models.CrossSlip.findAll({
+          where: where,
+          include: [
+            {
+              model: models.User,
+              as: 'creater'
+            }, {
+              model: models.User,
+              as: 'approver'
+            }, {
+              model: models.User,
+              as: 'updater'
+            }, {
+              model: models.CrossSlipDetail,
+              as: 'lines',
+            }
+          ],
+          order: [
+            [ 'year', 'ASC'],
+            [ 'month', 'ASC'],
+            [ 'day', 'ASC' ],
+            [ 'no', 'ASC' ],
+            ['lines', 'lineNo', 'ASC']
+          ]
+        });
+        res.json(cross_slips);
+        break;
+      default:
+        break;
+    }
+  },
+  get: async(req, res, next) => {
+    res.set('Access-Control-Allow-Origin', '*');
+    let year = req.params.year;
+    let month = req.params.month;
+    let no = req.params.no;
 
-		let cross_slip = await models.CrossSlip.findOne({
-			where: {
-				[Op.and]: {
-					year: year,
-					month: month,
-					no: no
-				}
-			},
-			include: [
-				{
-					model: models.CrossSlipDetail,
-					as: 'lines'
-				}
-			]
-		});
-		//console.log(cross_slip);
-		res.json(cross_slip);
-	},
-	post: async(req, res, next) => {
-		let body = req.body;
-		//console.log('body:', body);
-
-		fy = await models.FiscalYear.findOne({
-			where: {
-				startDate: {
-					[Op.lte]: new Date(body.year, body.month - 1, 2)
-				},
-				endDate: {
-					[Op.gte]: new Date(body.year, body.month - 1, 2)
-				}
-			}
-		});
-		//console.log('fy', fy);
-		ml = await models.MonthlyLog.findOne({
-			where: {
-				term: fy.term,
-				month: body.month
-			}
-		});
-		if	( !ml )	{
-			ml = await models.MonthlyLog.create({
-				term: fy.term,
-				month: body.month,
-				slipCount: 0,
-				voucharCount: 0
-			})
-		}
-		//console.log('ml', ml);
-		ml.slipCount += 1;
-
-		slip = await models.CrossSlip.create({
-			year: body.year,
-			month: body.month,
-			day: body.day,
-			no: ml.slipCount,
-			lineCount: body.lines.length,
-			term: body.term
-		});
-		//console.log(slip);
-		ml.save();
-
-		let lines = [];
-		for ( let i = 0; i < body.lines.length ; i ++ ) {
-			let line = body.lines[i];
-			line.crossSlipId = slip.id;
-			line.lineNo = i;
-			//console.log(line);
-			await models.CrossSlipDetail.create(line);
-			lines.push(line);
-		}
-		
-		res.json({
-			year: body.year,
-			month: body.month,
-			day: body.day,
-			no: slip.no,
-			lines: lines
-		});
-	},
-	update: async(req, res, next) => {
-		let body = req.body;
-		//console.log('update body:', body);
-		let slip = await models.CrossSlip.findOne({
-			where: {
-				year: body.year,
-				month: body.month,
-				no: body.no
-			}
-		});
-		if ( slip ) {
-			//console.log(slip);
-
-			slip.lineCount = body.lines.length;
-			slip.day = body.day;
-			slip.save();
-		
-			details = await models.CrossSlipDetail.findAll({
-				where: {
-					crossSlipId: slip.id
-				}
-			});
-			for ( let i = 0; i < details.length; i ++ ) {
-				//console.log('delete', details[i]);
-				await details[i].destroy();
-			}
-			for ( let i = 0; i < body.lines.length ; i ++ ) {
-				let line = body.lines[i];
-				line.crossSlipId = slip.id;
-				line.lineNo = i;
-				//console.log(line);
-				await models.CrossSlipDetail.create(line);
-			}
-			res.json({
-				code: 0
-			});
-		} else {
-			res.json({
-				code: -1,
-				message: 'record not found'
-			});
-		}
-
-	},
-	delete: async(req, res, next) => {
-		let body = req.body;
-		//console.log('body:', body);
-		let slip = await models.CrossSlip.findOne({
-			where: {
-				year: body.year,
-				month: body.month,
-				day: body.day,
-				no: body.no
-			}
-		});
-		//console.log('delete', slip);
-		details = await models.CrossSlipDetail.findAll({
-			where: {
-				crossSlipId: slip.id
-			}
-		});
-		for ( let i = 0; i < details.length; i ++ ) {
-			//console.log('delete', details[i]);
-			await details[i].destroy();
-		}
-		await slip.destroy();
-		res.json({
-			code: 0,
-		});
-	},
+    let cross_slip = await models.CrossSlip.findOne({
+      where: {
+        [Op.and]: {
+          year: year,
+          month: month,
+          no: no
+        }
+      },
+      include: [
+        {
+          model: models.CrossSlipDetail,
+          as: 'lines',
+          include: [
+            {
+              model: models.Voucher,
+              required: false,
+              as: 'debitVoucher',
+              include: [{
+                model: models.VoucherFile,
+                as: 'files'
+              }]
+            }, {
+              model: models.Voucher,
+              required: false,
+              as: 'creditVoucher',
+              include: [{
+                model: models.VoucherFile,
+                as: 'files'
+              }]
+            }, {
+              model: models.TaxRule,
+              as: 'debitTaxRule'
+            }, {
+              model: models.TaxRule,
+              as: 'creditTaxRule'
+            }
+          ]
+        },
+        {
+          model: models.User,
+          as: 'creater'
+        },
+        {
+          model: models.User,
+          as: 'approver'
+        }
+      ],
+      order: [
+        ['lines', 'lineNo', 'ASC']
+      ]
+    });
+    //console.log(cross_slip);
+    res.json(cross_slip);
+  },
+  post: async(req, res, next) => {
+    res.set('Access-Control-Allow-Origin', '*');
+    if	( req.session.user.accounting )  {
+      let body = req.body;
+      //console.log('body:', body);
+      let slip = await createCrossSlip(body, req.session.user);
+   
+      res.json(slip);
+    } else {
+      res.json({
+        code: -10,
+        message: 'this account can not create'
+      });
+    }
+  },
+  update: async(req, res, next) => {
+    let body = req.body;
+    let slip = await models.CrossSlip.findOne({
+        where: {
+          year: body.year,
+          month: body.month,
+          no: body.no
+        }
+      });
+    if ( slip ) {
+      if	( !slip.approvedAt )	{
+        if	(( req.session.user.accounting ) ||
+           ( req.session.user.id == slip.createdBy )) {
+            await updateCrossSlip(slip, body, req.session.user);
+            res.json({
+              code: 0
+            });
+        } else {
+        }
+      } else {
+        res.json({
+          code: -2,
+          message: 'this slip was approved'
+        });
+      }
+    } else {
+      res.json({
+        code: -1,
+        message: 'record not found'
+      });
+    }
+  },
+  delete: async(req, res, next) => {
+    res.set('Access-Control-Allow-Origin', '*');
+    if	( req.session.user.approvable )	{
+      let body = req.body;
+      //console.log('body:', body);
+      let slip = await models.CrossSlip.findOne({
+        where: {
+          year: body.year,
+          month: body.month,
+          day: body.day,
+          no: body.no
+        }
+      });
+      if	( !slip.approvedAt )	{
+        await slip.destroy();
+        res.json({
+          code: 0,
+        });
+      } else {
+        res.json({
+          code: -2,
+          message: 'thid slip was approved'
+        });
+      }
+    } else {
+      res.json({
+        code: -10,
+        message: 'this account can not delete'
+      });
+    }
+  },
+  approve: (req, res, next) => {
+    res.set('Access-Control-Allow-Origin', '*');
+    if	( req.session.user.approvable )	{
+      let body = req.body;
+      //console.log('update body:', body);
+      models.CrossSlip.findOne({
+        where: {
+          year: body.year,
+          month: body.month,
+          no: body.no
+        }
+      }).then((slip) => {
+        slip.approvedAt = body.approvedAt;
+        if	( body.approvedAt )	{
+          slip.approvedBy = req.session.user.id;
+        } else {
+          slip.approvedBy = null;
+        }
+        slip.updatedBy = req.session.user.id;
+        slip.save();
+        res.json({
+          code: 0,
+          id: slip.id
+        });
+      }).catch((e) => {
+        res.json({
+          code: -1,
+          message: 'record not found'
+        });
+      });
+    } else {
+      res.json({
+        code: -10,
+        message: 'this account can not approve'
+      });
+    }
+  }
 }
