@@ -1,170 +1,131 @@
-const element_index = (element) => {
-	return parseInt(element.id.match(/.*\[(\d+)\]/)[1]);
+import {numeric} from '../../libs/utils.js';
+
+export const element_index = (element) => {
+  return parseInt(element.id.match(/.*\[(\d+)\]/)[1]);
 }
-const element_dc = (element) => {
-	return element.id.split('-')[0];
+export const element_dc = (element) => {
+  return element.id.split('-')[0];
 }
 
 let accounts;
 
-const numeric = (s) => {
-	let ret;
-	let sign;
-
-	if ( s ) {
-		if ( typeof s == 'number' ) {
-			ret = s;
-		} else {
-			if ( s.length > 0 ) {
-				if ( s[0] == '-' ) {
-					sign = -1;
-				} else {
-					sign = 1;
-				}
-				let ss = s.replace(/[\D,\s]/g,'');
-				if ( ss.length > 0 ) {
-					ret = parseInt(ss) * sign;
-				} else {
-					ret = 0;
-				}
-			} else {
-			ret = 0;
-			}
-		}
-	} else {
-		ret = 0;
-	}
-	return ret;
+export const findAccount = (code) => {
+  //console.log(`account [${code}]`);
+  let account = { name: '', key: ''};
+  if ( accounts ) {
+    for ( let i = 0; i < accounts.length; i ++ ) {
+      if ( accounts[i].code == code ) {
+        account = accounts[i];
+        break;
+      }
+    }
+  }
+  return account;
+}
+export const findSubAccount = (account, code) => {
+  let sub_account = { name: '', key: ''};
+  
+  if	( ( account ) &&
+      ( account.subAccounts ) ) {
+    for ( let i = 0; i < account.subAccounts.length; i ++ ) {
+      if ( account.subAccounts[i].code == code ) {
+        sub_account = account.subAccounts[i];
+        break;
+      }
+    }
+  }
+  return sub_account;
 }
 
-const find_account = (code) => {
-	//console.log(`account [${code}]`);
-	let account = { name: '', key: ''};
-	if ( accounts ) {
-		for ( let i = 0; i < accounts.length; i ++ ) {
-			if ( accounts[i].code == code ) {
-				account = accounts[i];
-				break;
-			}
-		}
-	}
-	return account;
-}
-const find_sub_account = (account, code) => {
-	let sub_account = { name: '', key: ''};
-	
-	if	( ( account ) &&
-		  ( account.subAccounts ) ) {
-		for ( let i = 0; i < account.subAccounts.length; i ++ ) {
-			if ( account.subAccounts[i].code == code ) {
-				sub_account = account.subAccounts[i];
-				break;
-			}
-		}
-	}
-	return sub_account;
+export const findSubAccountByCode = (account_code, code) => {
+  let sub_account = { name: '', key: ''};
+
+  let account = findAccount(account_code);
+
+  if ( account.subAccounts ) {
+    for ( let i = 0; i < account.subAccounts.length; i ++ ) {
+      if ( account.subAccounts[i].code == code ) {
+        sub_account = account.subAccounts[i];
+        break;
+      }
+    }
+  }
+  return sub_account;
 }
 
-const find_sub_account_by_code = (account_code, code) => {
-	let sub_account = { name: '', key: ''};
-
-	let account = find_account(account_code);
-
-	if ( account.subAccounts ) {
-		for ( let i = 0; i < account.subAccounts.length; i ++ ) {
-			if ( account.subAccounts[i].code == code ) {
-				sub_account = account.subAccounts[i];
-				break;
-			}
-		}
-	}
-	return sub_account;
+export const findTaxClass = (ac, sub) => {
+  let tax = 0;
+  //console.log('findTaxClass', ac, sub);
+  for ( let i = 0; i < accounts.length; i ++ ) {
+    let account = accounts[i];
+    if ( account.code == ac ) {
+      if ( account.subAccounts ) {
+        for ( let j = 0 ; j < account.subAccounts.length; j ++ ) {
+          let sub_account = account.subAccounts[j];
+          if ( sub_account.code == sub ) {
+            tax = sub_account.taxClass;
+            break;
+          }
+        }
+      } else {
+        tax = account.taxClass;
+      }
+      break;
+    }
+  }
+  return tax;
 }
 
-const set_accounts = (arg) => {
-	accounts = arg;
+export const setAccounts = (arg) => {
+  accounts = arg;
 }
 
-const formatDate = (_date) => {
-	let date;
-	if ( _date )	{
-		date = new Date(_date);
-		return	`${date.getFullYear()}-${('0' + (date.getMonth()+1)).slice(-2)}-${('0' + date.getDate()).slice(-2)}`
-	} else {
-		return	('');
-	}
+export const invoiceStatus = (code) => {
+  if  ( code )  {
+    switch  (code)  {
+      case  11:
+        return  ('見積');
+        break;
+      case  21:
+        return  ('受注');
+        break;
+      case  31:
+        return  ('請求');
+        break;
+      case  32:
+        return  ('回収');
+        break;
+      default:
+        return  ('その他');
+        break;
+    }
+  }else{
+    return  ('未設定');
+  }
 }
 
-const voucherType = (code) => {
-	if	( code )	{
-		switch	(code)	{
-		  case	1:
-			return	('受取請求書');
-			break;
-		  case	2:
-			return	('受取領収書');
-			break;
-		  case	11:
-			return	('差出請求書');
-			break;
-		  case	12:
-			return	('差出領収書');
-			break;
-		  default:
-			return	('その他');
-			break;
-		}
-	} else {
-		return	('未設定');
-	}
+export const taxClass = (taxClass) => {
+  switch(taxClass) {
+    case 0:
+      return ("");
+    case 1:
+      return ('内税');
+    case 2:
+      return ('外税');
+    case 9:
+      return ('別計算');
+  }
+  return ('');
+}
+export default {
+  setAcounts: setAccounts,
+  findAcount: findAccount,
+  findSubAcount: findSubAccount,
+  findSuAcountByCode: findSubAccountByCode,
+  findTaxClass: findTaxClass,
+  element_index: element_index,
+  element_dc: element_dc,
+  invoiceStatus: invoiceStatus,
+  taxClass: taxClass,
 }
 
-const salesTax = (tax_class, _amount) => {
-	let amount = numeric(_amount);
-	let tax;
-	switch ( parseInt(tax_class) ) {
-	  case 1:
-		tax = Math.round(amount - amount / 110 * 100);
-		break;
-	  case 2:
-		tax = Math.round(amount * 0.1);
-		break;
-	  default:
-		tax = '';
-		break;
-	}
-	console.log('tax', tax);
-	return	(tax)
-}
-
-module.exports = {
-	set_accounts: set_accounts,
-	find_account: find_account,
-	find_sub_account: find_sub_account,
-	find_sub_account_by_code: find_sub_account_by_code,
-	element_index: element_index,
-	element_dc: element_dc,
-	numeric: numeric,
-	formatDate: formatDate,
-	voucherType: voucherType,
-	sales_tax: salesTax,
-	tax_class: (taxClass) => {
-		switch(taxClass) {
-			case 0:
-				return ("");
-			case 1:
-				return ('内税');
-			case 2:
-				return ('外税');
-			case 9:
-				return ('別計算');
-		}
-		return ('');
-	},
-	TAX_CLASS: [
-		[ '非課税', 0],
-		[ '内税',   1],
-		[ '外税',   2],
-		[ '別計算', 9]
-	]
-};
