@@ -1,2 +1,9 @@
-import 'bootstrap/dist/css/bootstrap.min.css';
 import 'bootstrap';
+
+import Login from '../svelte/login.svelte';
+
+const index = new Login({
+    target: document.getElementById('index'),
+});
+
+export default index;
