@@ -1,288 +1,282 @@
-<div class="d-flex justify-content-between mb-3 mt-3">
-  <h1 class="fs-3">銀行勘定帳</h1>
-</div>
-<div>
-  <ul class="navbar-nav me-auto mb-2">
+<div class="list">
+  <div class="page-title d-flex justify-content-between">
+    <h1>銀行元帳</h1>
+  </div>
+  <ul class="page-subtitle nav">
     <li class="nav-item dropdown">
-      <a class="btn btn-outline-primary dropdown-toggle" href="#"
+      <button type="button"
+        class="btn nav-link dropdown-toggle"
+        style="background-color:var(--bs-primary);color:white;"
         rolw="button" data-bs-toggle="dropdown" aria-expanded="false">
-        {#if this_account}
-        {BANK_ACCOUNTS.find((el) => el[0] == this_account)[1]}
+        {#if status.account}
+        {BANK_ACCOUNTS.find((el) => el[0] == status.account)[1]}
         {:else}
         科目
         {/if}
-      </a>
+      </button>
       <ul class="dropdown-menu" aria-labelledby="field">
         {#each BANK_ACCOUNTS as account}
         <li>
-          <a class="dropdown-item"
-            on:click={openAccount}
-            href="#"
-            data-account={account[0]}>{account[1]}</a>
+          <button type="button" class="btn btn-link dropdown-item"
+            on:click={() => {
+              openAccount(account[0])}
+            }>
+            {account[1]}
+          </button>
         </li>
         {/each}
       </ul>
     </li>
-  </ul>
-</div>
-{#if bank_list }
-  <div class="mb-2">
-    <ul class="nav">
-    {#each bank_list.SubAccounts as bank}
-      <li class="nav-item pe-2">
-        {#if ( sub_account == bank.id )}
-        <a class="btn btn-primary disabled"
-          on:click={openBank}
-          href="#"
-          data-account={this_account}
-          data-id={bank.id}>
+    {#each bank_list.subAccounts as bank}
+      <li class="nav-item">
+        {#if ( status.subAccount === bank.id )}
+        <button type="button" class="btn btn-info"
+          on:click|preventDefault={() => {
+            openBank(bank.subAccountCode);
+          }}>
           {bank.name}
-        </a>
+        </button>
         {:else}
-        <a class="btn btn-outline-primary"
-          on:click={openBank}
-          href="#"
-          data-account={this_account}
-          data-id={bank.id}>
+        <button type="button" class="btn btn-outline-info"
+          on:click|preventDefault={() => {
+            openBank(bank.subAccountCode);
+          }}>
           {bank.name}
-        </a>
+        </button>
         {/if}
       </li>
     {/each}
-    </ul>
+  </ul>
+  <div class="full-height" style="overflow-y: auto;">
+    <table class="table table-bordered">
+      <thead class="table-light">
+        <tr>
+          <th scope="col" colspan="2">
+            日付 / 伝番
+          </th>
+          <th scope="col" style="width: 150px;">
+            相手勘定科目<br/>相手補助科目
+          </th>
+          <th scope="col" style="width: 300px;">
+            適用<br/>補助科目
+          </th>
+          <th scope="col" style="width: 100px;">
+            支払金額
+          </th>
+          <th scope="col" style="width: 100px;">
+            預り金額
+          </th>
+          <th scope="col" style="width: 100px;">
+            残高
+          </th>
+        </tr>
+      </thead>
+      <tbody>
+      {#each lines as line}
+        <tr>
+          <td style="width:50px;text-align:center;">
+            {line.month} / {line.day}
+          </td>
+          <td style="width:50px;" class={'number ' + ( line.approvedAt ? 'bg-body' : 'bg-warning' )}>
+            <button type="button" class="btn btn-link"
+              on:click={() => {
+                openSlip(line.year, line.month, line.no)
+              }}>
+              {line.no}
+            </button>
+          </td>
+          <td>
+            {line.otherAccount}<br/>
+            {line.otherSubAccount}
+          </td>
+          <td>
+            <div class="application">
+              {line.application1 || ''}
+              {#if line.application2}
+              /
+              {line.application2}
+              {/if}
+            </div>
+            <div class="application d-flex">
+              <div class="tax">
+                {line.otherTaxRule}
+              </div>
+              <div class="">
+                {#if (line.debitVoucher )}
+                {#each line.debitVoucher.files as file}
+                <a href="/voucher/file/{file.id}" target="_blank">
+                  <i class="fas fa-file"></i>
+                </a>
+                {/each}
+                {/if}
+                {#if (line.creditVoucher )}
+                {#each line.creditVoucher.files as file}
+                <a href="/voucher/file/{file.id}" target="_blank">
+                  <i class="fas fa-file"></i>
+                </a>
+                {/each}
+                {/if}
+              </div>
+              <div class="ms-auto tax">
+                {line.thisTaxRule}
+              </div>
+            </div>
+          </td>
+          <td class="number">
+            {#if line.showCredit }
+            {line.pureCreditAmount ? line.pureCreditAmount.toLocaleString(): ''}
+            {/if}
+          </td>
+          <td class="number">
+            {#if line.showDebit }
+            {line.pureDebitAmount ? line.pureDebitAmount.toLocaleString(): ''}
+            {/if}
+          </td>
+          <td class="number">
+            {line.pureBalance.toLocaleString()}
+          </td>
+        </tr>
+      {/each}
+      </tbody>
+    </table>
   </div>
-{/if}
-<table class="table table-bordered">
-  <thead class="table-light">
-    <tr>
-      <th scope="col" colspan="2">
-        日付 / 伝番
-      </th>
-      <th scope="col" style="width: 150px;">
-        相手勘定科目<br/>相手補助科目
-      </th>
-      <th scope="col" style="width: 300px;">
-        適用<br/>補助科目
-      </th>
-      <th scope="col" style="width: 100px;">
-        支払金額
-      </th>
-      <th scope="col" style="width: 100px;">
-        預り金額
-      </th>
-      <th scope="col" style="width: 100px;">
-        残高
-      </th>
-    </tr>
-  </thead>
-  <tbody>
-  {#each lines as line}
-    <tr>
-      <td style="width:50px;text-align:center;">
-        {line.month} / {line.day}
-      </td>
-      <td class="number" style="width:50px;">
-        <a href="#" data-no={line.no} data-year={line.year} data-month={line.month}
-            on:click={openSlip}>
-          {line.no}
-        </a>
-      </td>
-      <td>
-        {line.otherAccount}<br/>
-        {line.otherSubAccount}
-      </td>
-      <td>
-        <div class="appication">
-          {line.application1 ? line.application1 : ''}
-        </div>
-        <div class="appication">
-          {line.application2 ? line.application2 : ''}
-        </div>
-      </td>
-      <td class="number">
-        {#if line.showCredit }
-        {line.pureCreditAmount ? line.pureCreditAmount.toLocaleString(): ''}
-        {/if}
-      </td>
-      <td class="number">
-        {#if line.showDebit }
-        {line.pureDebitAmount ? line.pureDebitAmount.toLocaleString(): ''}
-        {/if}
-      </td>
-      <td class="number">
-        {line.pureBalance.toLocaleString()}
-      </td>
-    </tr>
-  {/each}
-  </tbody>
-</table>
+</div>
+{#if popUp}
+{#key modalCount}
 <CrossSlipModal
-	slip={slip}
-	modal={modal}
-	term={term}
-	accounts={accounts}
-	on:close={updateList}></CrossSlipModal>
-
+  slip={slip}
+  status={status}
+  accounts={accounts}
+  bind:popUp={popUp}
+  on:close={updateList}></CrossSlipModal>
+{/key}
+{/if}
+  
 <style>
-.fontsize-12pt {
-	font-size: 12pt;
-}
-.application {
-	padding: 0;
-}
-th {
-	text-align: center;
-}
 </style>
 
 <script>
 import axios from 'axios';
-import Modal from 'bootstrap/js/dist/modal';
 import {onMount, beforeUpdate, afterUpdate, createEventDispatcher} from 'svelte';
-import {ledger_lines} from '../../../libs/ledger';
-import {set_accounts} from '../../javascripts/cross-slip';
+import {ledgerLines} from '../../../libs/ledger';
+import {setAccounts} from '../../javascripts/cross-slip';
 import CrossSlipModal from '../cross-slip/cross-slip-modal.svelte';
+import {DateString} from '../../../libs/utils.js';
 
-let	bank_list;
-let	term;
-let	this_account;
-let	sub_account;
-let	slip;
-let	lines;
-let	modal;
+export let status;
+
+let	bank_list = { subAccounts: []};
+let slip = {
+    year: 0,
+    month: 0,
+    lines: []
+  };
+let	lines = [];
 let	accounts;
+let modalCount = 0;
+let popUp;
+
 
 const BANK_ACCOUNTS = [
-	[ '1010000',	'当座預金' ],
-	[ '1010010',	'普通預金' ],
-	[ '1010020',	'定期預金' ],
-	[ '1010030',	'定期積立' ]
+  [ '1010000',	'当座預金' ],
+  [ '1010010',	'普通預金' ],
+  [ '1010020',	'定期預金' ],
+  [ '1010030',	'定期積立' ]
 ];
-const openAccount = (event) => {
-    event.preventDefault();
-	let dataset = event.currentTarget.dataset;
-	console.log('openBank');
-	this_account = dataset.account;
-	updateAccount(true);
+const openAccount = (account) => {
+  //console.log('openBank');
+  status.account = account;
+  updateAccount();
+  window.history.pushState(status, "",
+    `/bank-ledger/${status.account}`);
 }
-const openBank = (event) => {
-    event.preventDefault();
-	let dataset = event.currentTarget.dataset;
-	console.log('openAccount');
-	this_account = dataset.account;
-	sub_account = dataset.id;
-	updateList(true);
+const openBank = (id) => {
+  status.subAccount = id;
+  updateList();
+  window.history.pushState(status, "",
+    `/bank-ledger/${status.account}/${status.subAccount}`);
 }
-onMount(() => {
-	console.log('onMount');
+
+const update = () => {
+  updateAccount();
+  updateList();
+}
+const checkPage = () => {
+  update();
+}
+
+onMount(async () => {
+  console.log({status});
+  lines = [];
+  bank_list = { subAccounts: []};
+  let result = await axios.get('/api/accounts');
+  accounts = result.data;
+  setAccounts(accounts);
+  update();
 });
-beforeUpdate(()	=> {
-	let args = location.pathname.split('/');
-	if	( !term )	{
-		term = parseInt(args[2]);
-		this_account = args[3];
-		sub_account = args[4];
-		
-		window.onpopstate = (event) => {
-			if	( window.history.state )	{
-				let	state = window.history.state;
-				term = state.term;
-				this_account = state.account;
-				sub_account = state.sub_account;
-				lines = undefined;
-				bank_list = undefined;
-			}
-		}
-		if	( !accounts )	{
-			axios.get('/api/accounts').then((result) => {
-				accounts = result.data;
-				set_accounts(accounts);
-			});
-		}
-		console.log('beforeUpdate');
-	}
-	if	( !bank_list )	{
-		updateAccount();
-	}
-	if	( !lines)	{
-		lines = [];
-		updateList();
-	}
-	if	( !slip )	{
-		slip = {
-			year: 0,
-			month: 0,
-			lines: []
-		}
-	}
-});
-
-const updateAccount = (forward) => {
-	if	( this_account )	{
-		axios.get(`/api/account/${this_account}`).then((result) => {
-			bank_list = result.data;
-		});
-		if	( forward )	{
-				window.history.pushState({
-						term: term,
-						account: this_account
-					}, "", `/bank-ledget/${term}/${this_account}`);
-		}
-	}
-}
-
-const updateList = (forward) => {
-	if	( sub_account )	{
-		axios.get(`/api/remaining/${term}/${this_account}/${sub_account}`).then((result) => {
-			let remaining = result.data;
-			console.log('remaining', remaining);
-
-			axios.get(`/api/ledger/${term}/${this_account}/${sub_account}`).then((result) => {
-				let details = result.data;
-				let ret = ledger_lines(this_account, sub_account,
-												remaining, details);
-				lines = ret.lines;
-				console.log('lines', lines);
-			});
-			if	( forward )	{
-				window.history.pushState({
-						term: term,
-						account: this_account,
-						sub_account: sub_account
-					}, "", `/bank-ledger/${term}/${this_account}/${sub_account}`);
-			}
-		});
-	}
-
-}
-
-let openModal = false;
 afterUpdate(() => {
-	if	( !modal )	{
-		modal = new Modal(document.getElementById('cross-slip-modal'));
-		console.log('modal new')
-	}
-	if	( openModal )	{
-		modal.show();
-		openModal = false;
-	}
-});
+  if  (!popUp)  {
+    modalCount += 1;
+  }
+})
 
-const	openSlip = (event) => {
-	event.preventDefault();
-	let dataset = event.target.dataset;
-	console.log('openSlip', dataset);
-	axios.get(`/api/cross_slip/${dataset.year}/${dataset.month}/${dataset.no}`).then((result) => {
-		let data = result.data;
-		slip = {
-				year: data.year,
-				month: data.month,
-				day: data.day,
-				no: data.no,
-				lines: data.lines
-		};
-		openModal = true;
-		console.log('slip', slip);
-	});
+let _status;
+beforeUpdate(() => {
+  let args = location.pathname.split('/');
+  status.account = args[2];
+  status.subAccount = args[3] ? parseInt(args[3]) : undefined;
+  console.log('bank-ledger beforeUpdate', status);
+  if  (( status.change ) ||
+       ( _status !== status ))  {
+    status.change = false;
+    _status = status;
+    console.log('run checkPage');
+    checkPage();
+  }
+})
+
+const updateAccount = () => {
+  if	( status.account )	{
+    axios.get(`/api/account/${status.account}`).then((result) => {
+      bank_list = result.data;
+      console.log({bank_list});
+    });
+  }
+}
+
+const updateList = () => {
+  if	( status.subAccount )	{
+    axios.get(`/api/remaining/${status.fy.term}/${status.account}/${status.subAccount}`).then((result) => {
+      let remaining = result.data;
+      //console.log('remaining', remaining);
+
+      axios.get(`/api/ledger/${status.fy.term}/${status.account}/${status.subAccount}`).then((result) => {
+        let details = result.data;
+        console.log('details', details);
+        let ret = ledgerLines(status.account, status.subAccount,
+                        remaining, details);
+        lines = ret.lines;
+        console.log('lines', lines);
+      });
+    });
+  }
+}
+
+const	openSlip = (year, month, no) => {
+  axios.get(`/api/cross_slip/${year}/${month}/${no}`).then((result) => {
+    let data = result.data;
+    slip = {
+      year: data.year,
+      month: data.month,
+      day: data.day,
+      no: data.no,
+      createdBy: data.createdBy,
+      approvedAt: data.approvedAt ? new Date(data.approvedAt): null,
+      createrName: data.creater ? data.creater.name: '',
+      approverName: data.approver ? data.approver.name : '',
+      lines: data.lines
+    };
+    popUp = true;
+  });
 }
 </script>
