@@ -10,7 +10,6 @@ module.exports = {
 	mode: MODE,
 	context: path.resolve(__dirname, 'front/javascripts'),
     entry: {
-		login: './login.js',
 		home: './home.js',
 		common: './common.js',
 		journal: './journal.js',
@@ -20,7 +19,9 @@ module.exports = {
 		customer: './customer.js',
 		voucher: './voucher.js',
 		'bank-ledger': './bank-ledger.js',
-    setup: './setup.js'
+		index: './index.js',
+		changes: './changes.js',
+    	setup: './setup.js'
 	},
     output: {
         path: path.resolve(__dirname, 'dist'),
@@ -54,8 +55,38 @@ module.exports = {
 				]
 			},
 			{
-				test: /\.woff(2)?(\?v=[0-9]\.[0-9])?$/,
-				loader: "url-loader?limit=10000&mimetype=application/font-woff"
+				test: /\.m?js/,
+				resolve: {
+					fullySpecified: false
+				}
+			},
+			{
+				test: /\.(scss)$/,
+				use: [{
+					// Adds CSS to the DOM by injecting a `<style>` tag
+					loader: 'style-loader'
+				  },
+				  {
+					// Interprets `@import` and `url()` like `import/require()` and will resolve them
+					loader: 'css-loader'
+				  },
+				  {
+					// Loads a SASS/SCSS file and compiles it to CSS
+					loader: 'sass-loader',
+					options: {
+						  sassOptions: {
+							includePaths: ['node_modules']
+						  }
+					   }
+				}
+				]
+			  },
+			{
+				test: /\.(woff|woff2|eot|ttf|otf)$/i, 
+				type: 'asset/resource', 
+				generator: {
+					filename: 'fonts/[name][ext][query]'
+				}
 			},
             {
 				test: /\.svg(\?v=[0-9]\.[0-9]\.[0-9])?$/,
@@ -85,11 +116,15 @@ module.exports = {
 	resolve: {
 		alias: {
 			path: "path-browserify",
+			'chart.js$': path.resolve(__dirname, 'node_modules/chart.js/dist/Chart.js')
 		},
 		modules: [
 			'node_modules',
 			'front/javascripts',
 			'libs'
 		],
+		extensions: ['.mjs', '.js', '.svelte'],
+		mainFields: ['svelte', 'browser', 'module', 'main'],
+		conditionNames: ['svelte', 'browser']
 	}
 }
